de interleave to functions,  get focus to aux
